Added agent callbacks
Added the agent callbacks so that an agent can call back to
the neutron API server to retrieve information or perform
an action that an agent cannot do because it does not
have access to the same resources as the neutron API server
(such as DB access).

Also added from_dict methods in the data_models as callbacks
and the agent will need these.

Change-Id: I292415206b23731ea8197aeb91fc87ff148bc14c
Depends-On: Ic9179ef0a95b91d1b7662537fffeb0a949efc925
Partially-implements: blueprint lbaas-refactor-haproxy-namespace-driver-to-new-driver-interface
Co-Authored-By: Brandon Logan <brandon.logan@rackspace.com>