Finish Agent Scheduler and LB Agent Driver Piece
Implements the agent scheduler to choose which agent will host
the loadbalancer.  Only a random scheduling algorithm implemented.

Exposed two API calls that allows an admin user to get a list of
loadbalancers hosted by an agent and get the agent that is hosting
a particular loadbalancer.

Implemented the loadbalancer driver classes and methods for the agent
driver.  These will call the agent scheduler to host the load balancer
and then send the data to a queue.

Depends-On: Ic9179ef0a95b91d1b7662537fffeb0a949efc925
Partially-implements: blueprint lbaas-refactor-haproxy-namespace-driver-to-new-driver-interface
Co-Authored-By: Brandon Logan <brandon.logan@rackspace.com>

Change-Id: Ic4a83ef255aed0474eebb73f28d67e09f35666b7